Organization, Physics Update, Gestures, Level Gen
Updated some of the code organization.
Fixed some of the physics.
Added swipe gesture support for swiping left, right, up, and down. Then
implemented it in the "Store" activity.
Updated the level generator to support objects both on the ceiling and
in mid-air.
There may have been a few other changes...